Renamed files and removed unfinished documentation
