+++ Model/cell_methods.py
@@ -1,6 +1,6 @@
 import random as r
 
-from cellbackend import *
+from cell_backend import *
 from backend import record_time, normal_vector
