+++ Model/run.py
@@ -1,4 +1,4 @@
-from cellsimulation import CellSimulation
+from cell_simulation import CellSimulation
 
 # only call start() if this file is being run directly
 if __name__ == "__main__":
